Bohm Testing After Rescaling gamma factor
Made a units error in Bohm algorithm. Calling the density of the native LAMMPS units meant I was calling the density in amu / bohr^3. The density needs to be in m_e / bohr^3 to be in Hartree units. Correction has so far been made ad hoc via changing the input gamma_factor to Bohm loop, but could also be done internally in algorithm in future.